+++ tools/migrate/makefile
@@ -1,2 +1,5 @@
+TEST_FILES   := $(wildcard test_*.py)
+TEST_MODULES := $(TEST_FILES:%.py=%)
+
 test:
-	 PYTHONPATH=../.. python -m unittest test_operations
+	 PYTHONPATH=../.. python -m unittest $(TEST_MODULES)
